Move async helpers to AsyncEnumerableUtils; fix log message
Per review feedback from James Newton-King:
- Move ReadChannelUntilCancelledAsync and ReadEnumerableUntilCancelledAsync
  out of AppHostRpcTarget into new Aspire.Hosting.Utils.AsyncEnumerableUtils
  so they can be reused elsewhere
- Fix misleading log message that said 'due to AppHost shutdown' when the
  linked token also covers client-side cancellation
